+++ forge-game/src/main/java/forge/game/card/Card.java
@@ -1331,12 +1331,6 @@ public class Card extends GameEntity implements Comparable<Card>, IHasSVars, ITr
     public final void removeUntilLeavesBattlefield(final Card c) {
         untilLeavesBattlefield = view.removeCard(untilLeavesBattlefield, c, TrackableProperty.UntilLeavesBattlefield);
     }
-    public final void removeUntilLeavesBattlefield(final Iterable<Card> cards) {
-        untilLeavesBattlefield = view.removeCards(untilLeavesBattlefield, cards, TrackableProperty.UntilLeavesBattlefield);
-    }
-    public final void clearUntilLeavesBattlefield() {
-        untilLeavesBattlefield = view.clearCards(untilLeavesBattlefield, TrackableProperty.UntilLeavesBattlefield);
-    }
 
     public final CardCollectionView getExiledCards() {
         return CardCollection.getView(exiledCards);
@@ -1355,12 +1349,6 @@ public class Card extends GameEntity implements Comparable<Card>, IHasSVars, ITr
     }
     public final void removeExiledCard(final Card c) {
         exiledCards = view.removeCard(exiledCards, c, TrackableProperty.ExiledCards);
-    }
-    public final void removeExiledCards(final Iterable<Card> cards) {
-        exiledCards = view.removeCards(exiledCards, cards, TrackableProperty.ExiledCards);
-    }
-    public final void clearExiledCards() {
-        exiledCards = view.clearCards(exiledCards, TrackableProperty.ExiledCards);
     }
 
     public final CardCollectionView getHauntedBy() {
@@ -2561,17 +2549,10 @@ public class Card extends GameEntity implements Comparable<Card>, IHasSVars, ITr
                     sbLong.append("\r\n");
                 } else if (keyword.startsWith("Protection:")) {
                     final String[] k = keyword.split(":");
-                    sbLong.append("Protection from ");
                     if (k.length > 2) {
-                        sbLong.append(k[2]);
+                        sbLong.append("Protection from ").append(k[2]);
                     } else {
-                        if (MagicColor.Constant.ONLY_COLORS.contains(k[1])) {
-                            // lower-case color
-                            sbLong.append(k[1]);
-                        } else {
-                            // plural card types
-                            sbLong.append(CardType.getPluralType(k[1]));
-                        }
+                        sbLong.append(inst.getTitle());
                     }
                     sbLong.append("\r\n");
                 } else if (inst.getKeyword().equals(Keyword.COMPANION)) {
