+++ forge-game/src/main/java/forge/game/card/CardFactoryUtil.java
@@ -541,52 +541,6 @@ public class CardFactoryUtil {
         return re;
     }
 
-    public static String getProtectionValid(final String kw, final boolean damage) {
-        String validSource = "";
-
-        if (kw.startsWith("Protection:")) {
-            final String[] kws = kw.split(":");
-            String characteristic = kws[1];
-            if (characteristic.startsWith("Player")) {
-                validSource = "ControlledBy " + characteristic;
-            } else {
-                if (damage && (characteristic.endsWith("White") || characteristic.endsWith("Blue")
-                    || characteristic.endsWith("Black") || characteristic.endsWith("Red")
-                    || characteristic.endsWith("Green") || characteristic.endsWith("Colorless")
-                    || characteristic.endsWith("MonoColor") || characteristic.endsWith("MultiColor")
-                    || characteristic.endsWith("EnemyColor"))) {
-                    characteristic += "Source";
-                }
-                return characteristic;
-            }
-        } else if (kw.startsWith("Protection from ")) {
-            String protectType = kw.substring("Protection from ".length());
-            if (protectType.equals("white")) {
-                validSource = "White" + (damage ? "Source" : "");
-            } else if (protectType.equals("blue")) {
-                validSource = "Blue" + (damage ? "Source" : "");
-            } else if (protectType.equals("black")) {
-                validSource = "Black" + (damage ? "Source" : "");
-            } else if (protectType.equals("red")) {
-                validSource = "Red" + (damage ? "Source" : "");
-            } else if (protectType.equals("green")) {
-                validSource = "Green" + (damage ? "Source" : "");
-            } else if (protectType.equals("colorless")) {
-                validSource = "Colorless" + (damage ? "Source" : "");
-            } else if (protectType.equals("each color")) {
-                validSource = "nonColorless" + (damage ? "Source" : "");
-            } else if (protectType.equals("everything")) {
-                return "";
-            } else {
-                throw new RuntimeException("unknown protection keyword: " + kw);
-            }
-        }
-        if (validSource.isEmpty()) {
-            return validSource;
-        }
-        return "Card." + validSource + ",Emblem." + validSource;
-    }
-
     public static ReplacementEffect makeEtbCounter(final String kw, final CardState card, final boolean intrinsic) {
         String parse = kw;
 
@@ -2590,7 +2544,7 @@ public class CardFactoryUtil {
             }
         }
         else if (keyword.startsWith("Protection")) {
-            String validSource = getProtectionValid(keyword, true);
+            String validSource = Protection.getProtectionValid(keyword, true);
 
             String rep = "Event$ DamageDone | Prevent$ True | ActiveZones$ Battlefield | ValidTarget$ Card.Self";
             if (!validSource.isEmpty()) {
@@ -3964,7 +3918,7 @@ public class CardFactoryUtil {
             String effect = "Mode$ CantTransform | ValidCard$ Creature.Self | ExceptCause$ SpellAbility.Nightbound | Secondary$ True | Description$ This permanent can't be transformed except by its nightbound ability.";
             inst.addStaticAbility(StaticAbility.create(effect, state.getCard(), state, intrinsic));
         } else if (keyword.startsWith("Protection")) {
-            String valid = getProtectionValid(keyword, false);
+            String valid = Protection.getProtectionValid(keyword, false);
 
             // Block
             String effect = "Mode$ CantBlockBy | ValidAttacker$ Creature.Self | Secondary$ True ";
